Recaftor SVG icon generation methods into separate service.
